Add files via upload
What I am confused about is when the assignment states to add the other feature and mess around with k, does that mean add the column of duration to loudness so that the shape is (30,2) or does it mean instead to use duration instead of loudness?  And if the proper way to do it is to use the shape of (30,2), then what am I doing wrong? And why is my code not working out.  I understand that the shape of X and Y are different, but I have been racking my brain as to how I can change them so that Y_ = knn.predict(T) will work.